GeoFiltering UX Changes
- make "click" add polygons and "rightclick" remove them.  this makes it
  much nicer to zoom and navigate within a selected GeoJson feature.  Previously
  it was too easy to remove the polygon accidentally while panning around within
  a selected area.
- for bounding boxes make it consistent and have "rightclick" be the event that
  results in removal of a box.
- when drawing a new rectangle automatically reset the drawing control so you
  you don't have to go and change drawing modes.  This means the user can draw
  one rectangle at a time.  To draw another they must reselect the draw a rectangle
  control again.  In the end this is a less confusing UX.
- Since the drawing control is automatically reset hide the "hand" in the drawing
  control panel since it's no longer really necessary and frees up space (less for
  the user to worry about).
- move the drawing controls up into the right corner of the map and bump the
  filter tags a little left to give it room to be.

+++ src/js/layers/layers.js
@@ -276,10 +276,16 @@ angular.module('npn-viz-tool.layers',[
                 }
             });
 
+            function restyleAndRefilter() {
+                LayerService.restyleLayers().then(function(){
+                    if(FilterService.getFilter().hasSufficientCriteria()) {
+                        $rootScope.$broadcast('filter-rerun-phase2',{});
+                    }
+                });
+            }
+
             function clickFeature(feature,map) {
-                // TODO "NAME" may or may not be suitable, probably should use id...
-                var name = feature.getProperty('NAME'),
-                    filterArg = feature.getProperty('$FILTER');
+                var filterArg = feature.getProperty('$FILTER');
                 lastFeature = feature;
                 if(!filterArg) {
                     filterArg = new GeoFilterArg(feature,$scope.layerOnMap.layer.id);
@@ -287,17 +293,19 @@ angular.module('npn-viz-tool.layers',[
                     // TODO - different layers will probably have different styles, duplicating hard coded color...
                     // over-ride so the change shows up immediately and will be applied on the restyle (o/w there's a pause)
                     map.data.overrideStyle(feature, {fillColor: '#800000'});
-                } else {
-                    FilterService.removeFromFilter(filterArg);
-                    filterArg = null;
+                    feature.setProperty('$FILTER',filterArg);
+                    restyleAndRefilter();
                 }
-                feature.setProperty('$FILTER',filterArg);
-                LayerService.restyleLayers().then(function(){
-                    // TODO - maybe instead the filter should just broadcast the "end" event
-                    if(FilterService.getFilter().hasSufficientCriteria()) {
-                        $rootScope.$broadcast('filter-rerun-phase2',{});
-                    }
-                });
+            }
+
+            function rightClickFeature(feature,map) {
+                var filterArg = feature.getProperty('$FILTER');
+                lastFeature = feature;
+                if(filterArg) {
+                    FilterService.removeFromFilter(filterArg);
+                    feature.setProperty('$FILTER',null);
+                    restyleAndRefilter();
+                }
             }
 
 
@@ -367,7 +375,11 @@ angular.module('npn-viz-tool.layers',[
                             $scope.$apply(function(){
                                 clickFeature(event.feature,map);
                             });
-
+                        }));
+                        eventListeners.push(map.data.addListener('rightclick',function(event){
+                            $scope.$apply(function(){
+                                rightClickFeature(event.feature,map);
+                            });
                         }));
                     }
                     def.resolve(results);
